MediaItem and Album information are distributed in a separate json file
MediaItem and Album information are distributed in a separate json file.
User profile contains a new field 'mediaUri' containing the magnet-uri of the media/album json file.

This has been done in order to limit the size of the user profile file
because the value published in Vuze's DDB is apparently limited to 509
bytes.